Updates to piepline

+++ src/main.py
@@ -38,7 +38,9 @@ defaults = {
             "match_score": 1,
             "bias": 1.15,
             "output": "-",
-            "outsam": "-"
+            "outsam": "-",
+            "fq1": None,
+            "fq2": None
             }
 
 
@@ -61,8 +63,14 @@ def pipeline(kwargs):
 
     process, used_procs, remaining_procs = launch_external_mapper(kwargs)
     kwargs["procs"] = remaining_procs
-    kwargs["fq1"] = kwargs["out_pfix"] + "1.fq"
-    kwargs["fq2"] = kwargs["out_pfix"] + "2.fq"
+
+    if kwargs["mapper"] == "bwamem" and kwargs["map_script"] is None:
+        kwargs["fq1"] = None
+        kwargs["fq2"] = None
+    else:
+        kwargs["fq1"] = kwargs["out_pfix"] + "1.fq"
+        kwargs["fq2"] = kwargs["out_pfix"] + "2.fq"
+
     kwargs["sam"] = process.stdout
     kwargs["outsam"] = kwargs["out_pfix"] + ".sam"
 
@@ -71,8 +79,9 @@ def pipeline(kwargs):
     sort_and_index(kwargs)
 
     # Clean up
-    os.remove(kwargs["fq1"])
-    os.remove(kwargs["fq2"])
+    if kwargs["fq1"] is not None:
+        os.remove(kwargs["fq1"])
+        os.remove(kwargs["fq2"])
 
     if kwargs["mapper"] == "last":
         os.remove(kwargs["out_pfix"] + ".dict")
@@ -87,6 +96,45 @@ def pipeline(kwargs):
 
     click.echo("fufi run completed in {} h:m:s\n".format(str(datetime.timedelta(seconds=int(time.time() - t0)))),
                err=True)
+
+
+def launch_external_mapper(kwargs):
+    """Run a shell script to map interleaved .fastq files to .sam format. Uses a basic bwa-mem by default.
+    A custom shell script can be provided but must take positional arguments as:
+    $1 reference genome
+    $2 .fastq file; interleaved if paired-end, otherwise single end reads"""
+
+    if kwargs["procs"] > 1:
+        p = int(kwargs["procs"] / 2)
+        other_p = kwargs["procs"] - p
+    else:
+        p = kwargs["procs"]
+        other_p = 1
+
+    if not kwargs["map_script"] and kwargs["mapper"] == "bwamem":
+        command = "bwa mem -c 2000 -Y -t {procs} -P -a {ref} {s}1.fq {s}2.fq".format(procs=p,
+                                                                             ref=kwargs["reference"],
+                                                                             s=kwargs["fastq"])
+
+    elif not kwargs["map_script"] and kwargs["mapper"] == "last":
+        # Todo need to exract the sam header from input file, and use this as the dict argument in maf-convert
+        command = "fastq-interleave {s}1.fq {s}2.fq \
+        | lastal -k2 -l11 -Q1 -D10000 -K8 -C8 -i10M -r1 -q4 -a6 -b1 -P{procs} {ref} \
+        | last-map-probs -m 1 -s 1 | maf-convert -f {d}.dict sam".format(procs=p,
+                                              ref=kwargs["reference"],
+                                              d=kwargs["out_pfix"],
+                                              s=kwargs["fastq"])
+
+    else:
+        command = "bash {script} {ref} {s}1.fq {s}2.fq {procs}".format(script=kwargs["map_script"],
+                                                                       procs=p,
+                                                                       ref=kwargs["reference"],
+                                                                       s=kwargs["fastq"])
+
+    click.echo("Mapping command:\n" + command, err=True)
+    proc = Popen(command, stdout=PIPE, shell=True)
+
+    return proc, p, other_p
 
 
 def sort_and_index(kwargs):
@@ -153,45 +201,6 @@ def find_reads(**kwargs):
     return find_pairs.process(kwargs)
 
 
-def launch_external_mapper(kwargs):
-    """Run a shell script to map interleaved .fastq files to .sam format. Uses a basic bwa-mem by default.
-    A custom shell script can be provided but must take positional arguments as:
-    $1 reference genome
-    $2 .fastq file; interleaved if paired-end, otherwise single end reads"""
-
-    if kwargs["procs"] > 1:
-        p = int(kwargs["procs"] / 2)
-        other_p = kwargs["procs"] - p
-    else:
-        p = kwargs["procs"]
-        other_p = 1
-
-    if not kwargs["map_script"] and kwargs["mapper"] == "bwamem":
-        command = "bwa mem -t {procs} -P -a {ref} {s}1.fq {s}2.fq".format(procs=p,
-                                                                             ref=kwargs["reference"],
-                                                                             s=kwargs["fastq"])
-
-    elif not kwargs["map_script"] and kwargs["mapper"] == "last":
-        # Todo need to exract the sam header from input file, and use this as the dict argument in maf-convert
-        command = "fastq-interleave {s}1.fq {s}2.fq \
-        | lastal -k2 -l11 -Q1 -D10000 -K8 -C8 -i10M -r1 -q4 -a6 -b1 -P{procs} {ref} \
-        | last-map-probs -m 1 -s 1 | maf-convert -f {d}.dict sam".format(procs=p,
-                                              ref=kwargs["reference"],
-                                              d=kwargs["out_pfix"],
-                                              s=kwargs["fastq"])
-
-    else:
-        command = "bash {script} {ref} {s}1.fq {s}2.fq {procs}".format(script=kwargs["map_script"],
-                                                                       procs=p,
-                                                                       ref=kwargs["reference"],
-                                                                       s=kwargs["fastq"])
-
-    click.echo("Mapping command:\n" + command, err=True)
-    proc = Popen(command, stdout=PIPE, shell=True)
-
-    return proc, p, other_p
-
-
 @cli.command("align")
 @click.argument("sam", type=click.File('r'), required=True)
 @click.argument("output", required=False, type=click.Path())
@@ -199,9 +208,9 @@ def launch_external_mapper(kwargs):
 @click.option("--insert-stdev",  help="Template standard-deviation", default=defaults["insert_stdev"], type=float)
 @click.option("--read-length",  help="Length of a read in base-pairs", default=defaults["read_length"], type=float)
 @click.option("--fq1",  help="Fastq reads 1, used to add soft-clips to all hard-clipped read 1 alignments",
-              default=None, type=click.Path(exists=True))
+              default=defaults["fq1"], type=click.Path(exists=True))
 @click.option("--fq2",  help="Fastq reads 2, used to add soft-clips to all hard-clipped read 2 alignments",
-              default=None, type=click.Path(exists=True))
+              default=defaults["fq2"], type=click.Path(exists=True))
 @click.option("--max_insertion", help="Maximum insertion within read", default=defaults["max_insertion"], type=int)
 @click.option("--min-aln", help="Minimum alignment length", default=defaults["min_aln"], type=int)
 @click.option("--max-overlap", help="Maximum overlap between successive alignments", default=defaults["max_overlap"], type=int)
